+++ src/RuleSet/Php83.php
@@ -446,6 +446,7 @@ final class Php83
                 ],
                 'no_useless_else' => true,
                 'no_useless_nullsafe_operator' => true,
+                'no_useless_printf' => false,
                 'no_useless_return' => true,
                 'no_useless_sprintf' => true,
                 'no_whitespace_before_comma_in_array' => [