+++ src/RuleSet/Php84.php
@@ -448,6 +448,7 @@ final class Php84
                 ],
                 'no_useless_else' => true,
                 'no_useless_nullsafe_operator' => true,
+                'no_useless_printf' => false,
                 'no_useless_return' => true,
                 'no_useless_sprintf' => true,
                 'no_whitespace_before_comma_in_array' => [