+++ test/Unit/RuleSet/Php80Test.php
@@ -465,6 +465,7 @@ final class Php80Test extends ExplicitRuleSetTestCase
             ],
             'no_useless_else' => true,
             'no_useless_nullsafe_operator' => true,
+            'no_useless_printf' => false,
             'no_useless_return' => true,
             'no_useless_sprintf' => true,
             'no_whitespace_before_comma_in_array' => [