+++ src/RuleSet/Php54.php
@@ -424,7 +424,7 @@ final class Php54
                 ],
                 'no_useless_else' => true,
                 'no_useless_nullsafe_operator' => false,
-                'no_useless_printf' => false,
+                'no_useless_printf' => true,
                 'no_useless_return' => true,
                 'no_useless_sprintf' => true,
                 'no_whitespace_before_comma_in_array' => [