+++ src/RuleSet/Php81.php
@@ -439,7 +439,7 @@ final class Php81
                 ],
                 'no_useless_else' => true,
                 'no_useless_nullsafe_operator' => true,
-                'no_useless_printf' => false,
+                'no_useless_printf' => true,
                 'no_useless_return' => true,
                 'no_useless_sprintf' => true,
                 'no_whitespace_before_comma_in_array' => [