+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 For a full diff see [`6.49.0...main`][6.49.0...main].
 
+### Changed
+
+- Updated `friendsofphp/php-cs-fixer` ([#1249]), by [@dependabot]
+
 ## [`6.49.0`][6.49.0]
 
 For a full diff see [`6.48.0...6.49.0`][6.48.0...6.49.0].
@@ -1990,6 +1994,7 @@ For a full diff see [`d899e77...1.0.0`][d899e77...1.0.0].
 [#1243]: https://github.com/ergebnis/php-cs-fixer-config/pull/1243
 [#1244]: https://github.com/ergebnis/php-cs-fixer-config/pull/1244
 [#1245]: https://github.com/ergebnis/php-cs-fixer-config/pull/1245
+[#1249]: https://github.com/ergebnis/php-cs-fixer-config/pull/1249
 
 [@dependabot]: https://github.com/apps/dependabot
 [@linuxjuggler]: https://github.com/linuxjuggler