+++ CHANGELOG.md
@@ -11,6 +11,7 @@ For a full diff see [`6.49.0...main`][6.49.0...main].
 ### Changed
 
 - Updated `friendsofphp/php-cs-fixer` ([#1249]), by [@dependabot]
+- Enabled the `no_useless_printf` fixer ([#1252]), by [@localheinz]
 
 ### Fixed
 
@@ -1999,7 +2000,7 @@ For a full diff see [`d899e77...1.0.0`][d899e77...1.0.0].
 [#1244]: https://github.com/ergebnis/php-cs-fixer-config/pull/1244
 [#1245]: https://github.com/ergebnis/php-cs-fixer-config/pull/1245
 [#1249]: https://github.com/ergebnis/php-cs-fixer-config/pull/1249
-[#1251]: https://github.com/ergebnis/php-cs-fixer-config/pull/1251
+[#1252]: https://github.com/ergebnis/php-cs-fixer-config/pull/1252
 
 [@dependabot]: https://github.com/apps/dependabot
 [@linuxjuggler]: https://github.com/linuxjuggler