+++ CHANGELOG.md
@@ -6,7 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
-For a full diff see [`6.49.0...main`][6.49.0...main].
+For a full diff see [`6.50.0...main`][6.50.0...main].
+
+## [`6.50.0`][6.50.0]
+
+For a full diff see [`6.49.0...6.50.0`][6.49.0...6.50.0].
 
 ### Changed
 
@@ -1538,6 +1542,7 @@ For a full diff see [`d899e77...1.0.0`][d899e77...1.0.0].
 [6.47.0]: https://github.com/ergebnis/php-cs-fixer-config/releases/tag/6.47.0
 [6.48.0]: https://github.com/ergebnis/php-cs-fixer-config/releases/tag/6.48.0
 [6.49.0]: https://github.com/ergebnis/php-cs-fixer-config/releases/tag/6.49.0
+[6.50.0]: https://github.com/ergebnis/php-cs-fixer-config/releases/tag/6.50.0
 
 [d899e77...1.0.0]: https://github.com/ergebnis/php-cs-fixer-config/compare/d899e77...1.0.0
 [1.0.0...1.1.0]: https://github.com/ergebnis/php-cs-fixer-config/compare/1.0.0...1.1.0
@@ -1680,7 +1685,8 @@ For a full diff see [`d899e77...1.0.0`][d899e77...1.0.0].
 [6.46.0...6.47.0]: https://github.com/ergebnis/php-cs-fixer-config/compare/6.46.0...6.47.0
 [6.47.0...6.48.0]: https://github.com/ergebnis/php-cs-fixer-config/compare/6.47.0...6.48.0
 [6.48.0...6.49.0]: https://github.com/ergebnis/php-cs-fixer-config/compare/6.48.0...6.49.0
-[6.49.0...main]: https://github.com/ergebnis/php-cs-fixer-config/compare/6.49.0...main
+[6.49.0...6.50.0]: https://github.com/ergebnis/php-cs-fixer-config/compare/6.49.0...6.50.0
+[6.50.0...main]: https://github.com/ergebnis/php-cs-fixer-config/compare/6.50.0...main
 
 [#3]: https://github.com/ergebnis/php-cs-fixer-config/pull/3
 [#14]: https://github.com/ergebnis/php-cs-fixer-config/pull/14